ligando genero ao jogo, e jogo ao usuario

+++ retrocs/src/main/java/com/example/retrocs/repository/GeneroRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GeneroRepository extends JpaRepository<Genero, Integer> {}
+public interface GeneroRepository extends JpaRepository<Genero, Long> {}
